# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Flash Cards AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /words:
    get:
      operationId: getWords
      tags: [words]
      summary: List all words
      responses:
        "200":
          description: List of words
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Word"
    post:
      operationId: createWord
      tags: [words]
      summary: Create a word
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWordBody"
      responses:
        "201":
          description: Word created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Word"
  /words/{id}:
    get:
      operationId: getWord
      tags: [words]
      summary: Get a word by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Word found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Word"
    put:
      operationId: updateWord
      tags: [words]
      summary: Update a word
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWordBody"
      responses:
        "200":
          description: Word updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Word"
    delete:
      operationId: deleteWord
      tags: [words]
      summary: Delete a word
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /words/{id}/favorite:
    patch:
      operationId: toggleWordFavorite
      tags: [words]
      summary: Toggle word favorite status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Word"
...
components:
  schemas:
...
    Word:
      type: object
      properties:
        id:
          type: integer
        lessonId:
          type: integer
        word:
          type: string
        translation:
          type: string
        imageUrl:
          type: ["string", "null"]
        audioUrl:
          type: ["string", "null"]
        example:
          type: ["string", "null"]
        exampleTranslation:
          type: ["string", "null"]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        partOfSpeech:
          type: ["string", "null"]
        isFavorite:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - lessonId
        - word
        - translation
        - difficulty
        - isFavorite
        - createdAt
    CreateWordBody:
      type: object
      properties:
        lessonId:
          type: integer
        word:
          type: string
        translation:
          type: string
        imageUrl:
          type: ["string", "null"]
        audioUrl:
          type: ["string", "null"]
        example:
          type: ["string", "null"]
        exampleTranslation:
          type: ["string", "null"]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        partOfSpeech:
          type: ["string", "null"]
      required:
        - lessonId
        - word
        - translation
        - difficulty
...